# admin-panel/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shortlinker API",
    "description": "Shortlinker management API",
    "contact": {
      "name": "AptS-1547",
      "email": "apts-1547@esaps.net"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.5.2"
  },
  "paths": {
...
    "/admin/v1/config/reload": {
      "post": {
        "tags": [
          "config"
        ],
        "summary": "重新加载配置",
        "operationId": "reload_config",
        "responses": {
          "200": {
            "description": "Configuration reloaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReloadResponse"
                }
              }
            }
          },
          "500": {
            "description": "Configuration reload failed"
          }
        }
      }
    },
...
    "/admin/v1/config/{key}": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "获取单个配置",
        "operationId": "get_config",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Configuration key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ConfigItemResponse"
                }
              }
            }
          },
          "404": {
            "description": "Configuration key not found"
          }
        }
      },
      "put": {
        "tags": [
          "config"
        ],
        "summary": "更新配置",
        "operationId": "update_config",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Configuration key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Configuration updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ConfigUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid configuration value"
          },
          "404": {
            "description": "Configuration key not found"
          }
        }
      }
    },
    "/admin/v1/config/{key}/action": {
      "post": {
        "tags": [
          "config"
        ],
        "summary": "执行配置 action（如生成 token）",
        "description": "POST /admin/v1/config/{key}/action",
        "operationId": "execute_config_action",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Configuration key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Action result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ConfigActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unsupported action"
          }
        }
      }
    },
    "/admin/v1/config/{key}/execute-and-save": {
      "post": {
        "tags": [
          "config"
        ],
        "summary": "执行配置 action 并保存（安全版本）",
        "description": "POST /admin/v1/config/{key}/execute-and-save",
        "operationId": "execute_and_save_config_action",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Configuration key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Action executed and value saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ExecuteAndSaveResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unsupported action"
          }
        }
      }
    },
    "/admin/v1/config/{key}/history": {
      "get": {
        "tags": [
          "config"
        ],
        "summary": "获取配置变更历史",
        "operationId": "get_config_history",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Configuration key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_ConfigHistoryResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionType": {
        "type": "string",
        "description": "配置 Action 类型枚举\n\n用于标识配置项可执行的操作（如生成 token）。\n这是一个与 ValueType 正交的概念 - 任何类型的配置都可以有可选的 action。",
        "enum": [
          "generate_token"
        ]
      },
...
      "ApiResponse_ConfigActionResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "配置 action 执行响应",
            "required": [
              "value"
            ],
            "properties": {
              "value": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiResponse_ConfigItemResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "配置项响应",
            "required": [
              "key",
              "value",
              "value_type",
              "requires_restart",
              "is_sensitive",
              "updated_at"
            ],
            "properties": {
              "is_sensitive": {
                "type": "boolean"
              },
              "key": {
                "type": "string"
              },
              "requires_restart": {
                "type": "boolean"
              },
              "updated_at": {
                "type": "string"
              },
              "value": {
                "type": "string"
              },
              "value_type": {
                "$ref": "#/components/schemas/ValueType",
                "description": "值类型"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiResponse_ConfigUpdateResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "配置更新响应",
            "required": [
              "key",
              "value",
              "requires_restart",
              "is_sensitive",
              "message"
            ],
            "properties": {
              "is_sensitive": {
                "type": "boolean"
              },
              "key": {
                "type": "string"
              },
              "message": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "requires_restart": {
                "type": "boolean"
              },
              "value": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponse_ExecuteAndSaveResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "执行并保存响应（安全版本，不返回密钥值）",
            "required": [
              "success",
              "requires_restart",
              "message"
            ],
            "properties": {
              "message": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "requires_restart": {
                "type": "boolean"
              },
              "success": {
                "type": "boolean"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponse_ReloadResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "Reload 成功响应",
            "required": [
              "message",
              "duration_ms"
            ],
            "properties": {
              "duration_ms": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "message": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponse_Vec_ConfigHistoryResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "配置历史记录响应",
              "required": [
                "id",
                "config_key",
                "old_value",
                "new_value",
                "changed_at",
                "changed_by"
              ],
              "properties": {
                "changed_at": {
                  "type": "string"
                },
                "changed_by": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "config_key": {
                  "type": "string"
                },
                "id": {
                  "type": "integer",
                  "format": "int32"
                },
                "new_value": {
                  "type": "string"
                },
                "old_value": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ConfigActionRequest": {
        "type": "object",
        "description": "配置 action 执行请求",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionType"
          }
        }
      },
...
      "ConfigUpdateRequest": {
        "type": "object",
        "description": "配置更新请求",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
...
      "ValueType": {
        "type": "string",
        "description": "配置值类型枚举\n\n用于标识配置项在数据库和前端的类型。",
        "enum": [
          "string",
          "int",
          "float",
          "bool",
          "json",
          "enum",
          "stringarray",
          "enumarray"
        ]
      }
    }
  },
...
}